requirements.txt
streamlit==1.41.1
mysql-connector-python==9.1.0
streamlit_option_menu==0.4.0